create modules, transform functions to arrow functions, add docs
